rest

+++ restBank/bank/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Bank(models.Model):
     accno=models.IntegerField()
     username=models.CharField(max_length=20)
-    balance=models.IntegerField(default=0)
+    balance=models.FloatField(default=0)
 
     def __str__(self):
         return self.username
